Checkpoint: move AVFoundation responsibilities into UrgeSaver
